move uptake_model into separate module
